AI/requirements.txt
absl-py==2.2.2
astunparse==1.6.3
certifi==2025.1.31
charset-normalizer==3.4.1
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
keras==3.9.2
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.0.8
numpy==1.26.4
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
protobuf==4.25.6
Pygments==2.19.1
requests==2.32.3
rich==14.0.0
six==1.17.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
asttokens==3.0.0
attrs==25.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.3.2
debugpy==1.8.13
decorator==5.2.1
dill==0.3.8
etils[epath]==1.12.1
executing==2.2.0
filelock==3.17.0
Flask==3.0.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec>=2023.5.0,<2025.0.0
huggingface-hub==0.29.3
idna==3.10
imbalanced-learn==0.13.0
importlib_resources==6.5.2
ipykernel==6.29.5
jax==0.5.2
jaxlib==0.5.1
jedi==0.19.2
joblib==1.4.2
jsonpath-rw==1.4.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
matplotlib==3.10.1
matplotlib-inline==0.1.7
mediapipe==0.10.21
ml_dtypes==0.5.1
mlcroissant==1.0.14
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
opencv-contrib-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
pandas-stubs==2.2.3.250308
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
ply==3.11
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
rdflib==7.1.3
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
six==1.17.0
sklearn-compat==0.1.3
sounddevice==0.5.1
stack-data==0.6.3
threadpoolctl==3.5.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-pytz==2025.2.0.20250326
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0